Rename pure fixture dep
