+++ packages/fixture-dependency-pure/index.cjs.js
