+++ packages/fixture-dependency-pure/index.es.js
