+++ packages/fixture-dependency-pure/package.json
