+++ packages/fixture-package/flow-typed/cup.js
@@ -4,6 +4,6 @@ declare var __BROWSER__: boolean;
 declare var __NODE__: boolean;
 declare var __DEV__: boolean;
 
-declare module '@cup/fixture-pure-dependency' {
+declare module '@cup/fixture-dependency-pure' {
   declare export default any
 }
