+++ packages/fixture-package/src/index.js
@@ -2,7 +2,7 @@
 
 import a from './foo/a.js';
 import b from './foo/b.js';
-import pure from '@cup/fixture-pure-dependency';
+import pure from '@cup/fixture-dependency-pure';
 import React from 'react';
 
 import crypto from 'crypto';
